add type scriptness

+++ src/layer.c
@@ -5,7 +5,7 @@
 Layer *layer_create(layer_params params) {
   Layer *layer = nm_malloc(params.nm, sizeof(Layer));
   layer->nm = params.nm;
-  layer->neurons = nm_malloc(params.nm, sizeof(Neuron *) * params.nout);
+  layer->neurons = nm_malloc(params.nm, sizeof(Neuron *) * (size_t)params.nout);
   for (int i = 0; i < params.nout; i++) {
     layer->neurons[i] = neuron_create((neuron_params){
         .nin = params.nin,
@@ -31,7 +31,8 @@ int layer_nparams(Layer *layer) {
 }
 
 Value **layer_parameters(Layer *layer) {
-  Value **params = nm_calloc(layer->nm, layer_nparams(layer), sizeof(Value *));
+  Value **params =
+      nm_calloc(layer->nm, (size_t)layer_nparams(layer), sizeof(Value *));
   int idx = 0;
   for (int i = 0; i < layer->nout; i++) {
     Value **neuron_params = neuron_parameters(layer->neurons[i]);
